requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
requests==2.32.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asyncpg==0.30.0 